scan_server/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyMuPDF
PyYAML
pytesseract
Pillow
